chore(simplify): dedupe TIFF/IFD walker + brand-parser MF check
/simplify pass on top of the 8-commit feature work:

- Extract TIFFIFDReader.swift with ByteOrder + locateTIFFStart +
  findEntry / findTagValueField. EXIFOffsetReader and
  SonyMakerNoteReader both walk the same TIFF→IFD0 path; both used
  to carry their own copy of these helpers (the latter even called
  out the duplication in a comment). Now both consume the shared
  helper.
- Collapse the 5x repeated FocusMode MF/MANUAL check across Sony /
  Canon / Olympus / Fujifilm / Panasonic parsers into a single
  isManualFocus(_:extraSentinels:) private helper. Sony's numeric
  "1" sentinel is passed via extraSentinels.
- Drop parseOlympusFocusPoint's unused imageWidth/imageHeight
  parameters and the dead "_ = (imageWidth, imageHeight)" placeholder.

Net: -176 lines, no behavior change. 586/586 L1 tests pass.
